add initial working solution of find_phrases
This solution is quite naive in that it simply explores every possible permutation of phrases, from shortest to longest, and checks whether they are legal, yielding them if they are.

In particular, it rapidly gets very slow, even just for three-word phrases.

In practice, this is "good enough" to get a working, optimal solution to the puzzle. But it isn't very satisfying, and it seems there should be a smarter way to do this.